backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-allauth==0.53.0
django-cors-headers==4.1.0
django-axes==7.0.2
django-csp==4.0
django-otp==1.6.0
Pillow>=11.1.0
psycopg2-binary>=2.9.6  # Let pip find latest compatible version
python-dotenv==1.0.0
django-environ==0.9.0
django-filter==23.2
# drf-yasg==1.20.0  # Remove deprecated package